requirements.txt
vk_api
vk-api-bot-longpoll
pandas
gspread
oauth2client
python-dotenv